lib: adding web components and adjusting types
